+++ tests/testthat/test-add_gloss_gloss_example_make_gloss_list.R
@@ -8,6 +8,7 @@ test_that("add_gloss(), gloss_example(), make_gloss_list() test", {
                     output_format = "html_document")
   # rmarkdown::render(paste0(tmp, "/skeleton.Rmd"), quiet = TRUE,
   #                   output_format = "pdf_document")
-  expect_true(file.exists(paste0(tmp, "/skeleton.html")))
+  expect_true(file.exists(paste0(tmp, "/skeleton.html")),
+              label = "something get wrong in html version")
   # expect_true(file.exists(paste0(tmp, "/skeleton.pdf")))
 })