+++ NEWS.md
@@ -11,4 +11,6 @@
 - update `lingglosses::glosses_df` file with glosses from Wikipedia
 - make the package work with `bookdown`
 - add lingglosses rmarkdown template
+- add possibility to remove glosses in the `make_gloss_list()` output with the `remove_glosses` argument.
+- add possibility to escape glossing with curly brackets, e. g. `{I}`.